+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## [0.0.1] - Initial commit of working code
 
-* Working code for String.replace extension functionality
+* Working code for String.replace extension functionality and the humioManager